refactor(providers): share openai stream normalization
Move OpenAI-compatible SSE normalization into the neutral streaming module so Bedrock and OpenAI-compatible providers share protocol infrastructure without depending on each other.
